+++ composeApp/src/commonMain/kotlin/ru/nsu/teamsoul/data/repository/GameRepository.kt
@@ -6,7 +6,7 @@ interface GameRepository {
     suspend fun getGamesList(): Result<List<GamePluginsResponse>>
     suspend fun createGameInstance(gameId: Long, roomId: Long): Result<String>
 
-    suspend fun createRoom(): Result<Int>
+    suspend fun createRoom(): Result<Long>
     suspend fun joinRoom(roomId: String): Result<String>
-    suspend fun createGameInRoom(gameId: Int, roomId: Int): Result<String>
+    suspend fun createGameInRoom(gameId: Long, roomId: Long): Result<String>
 }
